Simplify zone-related code.
We no longer have specific configuration for each "zone"
of the board.  Instead, we infer the locations of a zone
on the board using the position of zone_color in the list
of zone_colors.

For example, if the zone_colors were ["blue", "red", "yellow"],
we'd figure out dynamically how much to rotate each color:

    blue - 0
    red - 120 degress
    yellow - 240 degrees

We also establish the convention that the first color in
zone_colors is the active color, so there will be a tiny
bit less of housekeeping to do when we start allowing
multiple players to take turns (and thus rotate the board
by rotation the elements of zone_colors).