Replenish hand if player uses all 5 cards.
In some situations a player has 5 face cards
(including jokers), and they have to play
all 5 cards then draw a completely new
hand.  The game engine now does this
automatically.  (We need to be a bit careful
not to "deal" the new cards until the user
actually plays the fifth card.)