refactor: Change Move to be a tuple.
We now have Move be a tuple with:

    move_type
    start_loc
    end_loc

This will allow us to unify Move and
CardStartEnd in subsequent commits.

For now I have to introduce the hacky
`Force` value in MoveType, but I can
eliminate that once I do a better job
of persisting move types on the player.turn
data.