Kill off NewSeed/randomize misfeature.
I'm not super concerned about players cheating by
reverse engineering the progression of random numbers.

Anybody smart enough to do that will beat 99% of
players anyhow, and the game is supposed to be fun.

Also, less randomizing will make it easier to have
undo work (and close a different kind of backdoor
for cheating, where you undo to get better cards).

I may later decide to attach the seed to individual
players, but it's low on my list.