Get move_type from moves in TurnNeedEndLoc for jacks.
We now get move_type from the moves on the Turn data
for a player when we're ready to execute the move,
rather than looking at the play_type.

For most moves, once we have the end location (from
the click), we don't really care about the move type,
or even the play type for that matter.

But for jack/seven moves, it's helpful to look at
the list of valid moves that we originally calculated,
so that we know how far we went (for 7 splits) or why
we were able to move there (for J trades).

In this commit, we make our handling of Jack trades
more clean, since the refactor forced our hand here
a bit.

For sevens, we'll do a follow up commit.