Simplify how we split sevens.
Before this change, if a user split a seven, we would
do an expensive/complex computation of the distance
between start_loc and end_loc for the first move, to
figure out how many moves we have left.

After this change, we get the distance from the move_type
that we already had from our original computation of
possible moves.

We will actually nuke the distance function in a
subsequent commit.