Prevent entering bullseye on certain seven splits.
We don't allow users to enter the bullseye on the first
half of a seven split.  (This rule is sort of controversial,
but I am adhering to the rules I learned, where you can't
enter the bullseye on the first piece you move during a
seven split.  Basically, the rationale is that you must
be in the bullseye at the end of playing a card, kinda
like you can't move through the bullseye with just one
card played.)

This also fixes a broken test.  (There were two broken
tests, where one just required fixing the test, and the
other required making the rules more strict.)