refactor: Stop passing around activeColor so much.
We now have activeColor directly on the model, and
functions that already have the model just pull it
off the model as needed.

(We'll kill off more code by putting activePlayer on
the model, too.)

The only place we change the activeColor is when
we rotate the board, so it's easy to isolate that
complexity.