major: Extract Game.elm, Game type, GameMsg.
This is a major restructuring.  We want to decouple
Game from the Model, so that we can handle
non-game situations like getting a random seed
without cluttering the code.  This may also be
helpful for having multiple games or implementing
undo/redo.

This also includes a few minor fixes.

I also rename ft.elm -> Main.elm.