major refactor: Make color an attribute of Player.
Now we just key players by a player index (i.e an
Int), and we store color directly inside the
Player object.  This removes some confusing code
where we infer the color of the player from a
rotated list of zone colors.

Now we just have a constant list of zone colors,
which doesn't change as we shift players.  Instead
we just bump the player index when a full player
turn is finished.

This already decouples things a bit.  If we wanted
to, we could keep the view always showing the same
board orientation.  Or, as I prefer now, we can
show a rotated view of the board, but that's now
completely controlled by the view, rather than
model updates.

I feel like I already understand my own code better,
but this will be particularly nice when I start
implementing partner play and/or allow players
to control multiple pieces.

Yet another benefit of this refactor is that it
paves the way to represent zone colors as a custom
type, rather than a string.  Right now I have
this hacky notion that the bullseye is in the
"black" zone, where really it's kinda special
in some ways.  But custom types can't be the
keys of dictionaries.  (Of course, I could have
just switched PlayerDict to an AssocList, too,
but it's nice to have fewer design constraints.)